Add script that read out sdc:nic ip information to generate nsd_listen.conf file
If we could read the information from sdc:nic and parse it we listen on
that ip addresses. If not the nsd_listen.conf is cleaned to we try to
listen on all interface (which moslty result in the bug).

We're also sure we listen on localhost ipv4 and ipv6.